+++ Johnson/docs/release-notes.md
@@ -2,6 +2,12 @@
 
 # Release notes
 
+## Upcoming release
+
+* Rebalanced Profit Margin.
+* Rebalanced Layout/Strategize-generated cards.
+* Reverted the "Layout and Strategize now put the cards in the discard pile instead of the draw pile." change.
+
 ## 1.0.3
 Released 13 February 2024.